+++ x/marker/keeper/keeper_test.go
@@ -64,7 +64,13 @@ func TestAccountMapperGetSet(t *testing.T) {
 	require.True(t, ok)
 	require.True(t, mac.AddressHasAccess(user, types.Access_Admin))
 
+	// add something to the send deny list just to verify removal
+	app.MarkerKeeper.AddSendDeny(ctx, addr, addr)
+
 	app.MarkerKeeper.RemoveMarker(ctx, mac)
+
+	// marker should not exist in send deny list
+	require.Empty(t, app.MarkerKeeper.GetSendDenyList(ctx, addr), "should not have entries in send deny list")
 
 	// getting account after delete should be nil
 	acc = app.AccountKeeper.GetAccount(ctx, addr)
@@ -1314,6 +1320,272 @@ func TestMsgUpdateRequiredAttributesRequest(t *testing.T) {
 func TestGetAuthority(t *testing.T) {
 	app := simapp.Setup(t)
 	require.Equal(t, "cosmos10d07y265gmmuvt4z0w9aw880jnsr700j6zn9kn", app.MarkerKeeper.GetAuthority())
+}
+
+func TestClearSendDeny(t *testing.T) {
+	app := simapp.Setup(t)
+	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
+
+	type SendDenyPair struct {
+		marker   sdk.AccAddress
+		sendDeny sdk.AccAddress
+	}
+
+	tests := []struct {
+		name   string
+		pairs  []SendDenyPair
+		marker sdk.AccAddress
+	}{
+		{
+			name:   "non existant marker",
+			pairs:  []SendDenyPair{},
+			marker: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+		},
+		{
+			name: "single entry for marker",
+			pairs: []SendDenyPair{
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+			},
+			marker: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+		},
+		{
+			name: "multiple entry for marker",
+			pairs: []SendDenyPair{
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+				},
+			},
+			marker: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+		},
+		{
+			name: "multiple markers with multiple entries",
+			pairs: []SendDenyPair{
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+				},
+				{
+					marker:   sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+				{
+					marker:   sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+					sendDeny: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+				},
+			},
+			marker: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			for _, pair := range tc.pairs {
+				app.MarkerKeeper.AddSendDeny(ctx, pair.marker, pair.sendDeny)
+			}
+
+			app.MarkerKeeper.ClearSendDeny(ctx, tc.marker)
+			list := app.MarkerKeeper.GetSendDenyList(ctx, tc.marker)
+			assert.Empty(t, list, "should remove all entries from send deny list")
+
+			for _, pair := range tc.pairs {
+				app.MarkerKeeper.RemoveSendDeny(ctx, pair.marker, pair.sendDeny)
+			}
+		})
+	}
+}
+
+func TestGetSendDenyList(t *testing.T) {
+	app := simapp.Setup(t)
+	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
+
+	type SendDenyPair struct {
+		marker   sdk.AccAddress
+		sendDeny sdk.AccAddress
+	}
+
+	tests := []struct {
+		name     string
+		pairs    []SendDenyPair
+		marker   sdk.AccAddress
+		expected []sdk.AccAddress
+	}{
+		{
+			name:     "non existant marker",
+			pairs:    []SendDenyPair{},
+			marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+			expected: []sdk.AccAddress{},
+		},
+		{
+			name: "single entry for marker",
+			pairs: []SendDenyPair{
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+			},
+			marker: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+			expected: []sdk.AccAddress{
+				sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+			},
+		},
+		{
+			name: "multiple entry for marker",
+			pairs: []SendDenyPair{
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+				},
+			},
+			marker: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+			expected: []sdk.AccAddress{
+				sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+				sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+			},
+		},
+		{
+			name: "multiple markers with multiple entries",
+			pairs: []SendDenyPair{
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+				},
+				{
+					marker:   sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+				{
+					marker:   sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+					sendDeny: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+				},
+			},
+			marker: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+			expected: []sdk.AccAddress{
+				sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+				sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			for _, pair := range tc.pairs {
+				app.MarkerKeeper.AddSendDeny(ctx, pair.marker, pair.sendDeny)
+			}
+
+			list := app.MarkerKeeper.GetSendDenyList(ctx, tc.marker)
+			assert.Equal(t, tc.expected, list, "should return the correct send deny entries for a marker")
+
+			for _, pair := range tc.pairs {
+				app.MarkerKeeper.RemoveSendDeny(ctx, pair.marker, pair.sendDeny)
+			}
+		})
+	}
+}
+
+func TestAddRemoveSendDeny(t *testing.T) {
+	app := simapp.Setup(t)
+	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
+
+	type SendDenyPair struct {
+		marker   sdk.AccAddress
+		sendDeny sdk.AccAddress
+	}
+
+	tests := []struct {
+		name  string
+		pairs []SendDenyPair
+	}{
+		{
+			name: "new marker and send deny pair",
+			pairs: []SendDenyPair{
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+			},
+		},
+		{
+			name: "duplicate marker and send deny pair",
+			pairs: []SendDenyPair{
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+			},
+		},
+		{
+			name: "multiple senders for marker",
+			pairs: []SendDenyPair{
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+				},
+			},
+		},
+		{
+			name: "multiple markers",
+			pairs: []SendDenyPair{
+				{
+					marker:   sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+					sendDeny: sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+				},
+				{
+					marker:   sdk.AccAddress("cosmos1w6t0l7z0yerj49ehnqwqaayxqpe3u7e23edgma"),
+					sendDeny: sdk.AccAddress("cosmos1v57fx2l2rt6ehujuu99u2fw05779m5e2ux4z2h"),
+				},
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			for _, pair := range tc.pairs {
+				app.MarkerKeeper.AddSendDeny(ctx, pair.marker, pair.sendDeny)
+			}
+
+			for _, pair := range tc.pairs {
+				isSendDeny := app.MarkerKeeper.IsSendDeny(ctx, pair.marker, pair.sendDeny)
+				require.True(t, isSendDeny, "should have entry for added pair")
+			}
+
+			for _, pair := range tc.pairs {
+				app.MarkerKeeper.RemoveSendDeny(ctx, pair.marker, pair.sendDeny)
+			}
+
+			for _, pair := range tc.pairs {
+				isSendDeny := app.MarkerKeeper.IsSendDeny(ctx, pair.marker, pair.sendDeny)
+				require.False(t, isSendDeny, "should not have entry for removed pair")
+			}
+		})
+	}
 }
 
 func TestReqAttrBypassAddrs(t *testing.T) {